kafka-producer/requirements.txt
kafka-python==2.0.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3 